+++ jwskate/jwt/signed.py
@@ -237,36 +237,6 @@ class SignedJwt(Jwt):
             return jti
         raise AttributeError("jti has an unexpected type", type(jti))
 
-    @cached_property
-    def alg(self) -> Optional[str]:
-        """Get the signature algorithm from the header.
-
-        Returns:
-          the token signing alg, from the `alg` header
-
-        Raises:
-            AttributeError: if the alg is not a string
-        """
-        alg = self.get_header("alg")
-        if alg is None or isinstance(alg, str):
-            return alg
-        raise AttributeError("alg has an unexpected type", type(alg))
-
-    @cached_property
-    def kid(self) -> Optional[str]:
-        """Get the Key ID (kid) from the JWT header.
-
-        Returns:
-          the token signing key id, from the `kid` header
-
-        Raises:
-            AttributeError: if the kid is present but is not a string
-        """
-        kid = self.get_header("kid")
-        if kid is None or isinstance(kid, str):
-            return kid
-        raise AttributeError("kid has an unexpected type", type(kid))
-
     def get_claim(self, key: str, default: Any = None) -> Any:
         """Get a claim from this Jwt.
